database stuff
Getting the database and library management code started. I'm just getting things to a base level where I can run the crawler and generate a good database to aid in further development. Its more difficult to make a function that displays database information when you don't have that database in the first place.