Folder scanning
Got started on the folder scanning. It mostly ignores TV shows with the regular expressions I've implemented. I've noticed a few files I have are incomprehensible filenames with normal folder names so I may have to take the folder name instead for better accuracy. A few files don't even have a proper folder name either so that's fun. We'll address that specific issue much later.

For now the next step is to isolate the movie title and then pull its data using the IMDBpy module, then save that data to the database.