Progress bar issues
Added a progress bar for the scan operation and ran into some troubles. When you use a statusbar you need a rough idea of how many operations you need to perform and what operation number you are currently on. That way you can give a percentage of the progress.

The way the scanner was designed it spins off multiple threads that first discover all dirs and then build up a file list. Because each thread encounters different numbers of files, one thread will always finish enumerating directories before the others. When this thread finishes enumerating dirs it then starts adding files to the database.

The issue is this: We want to be able to give accurate info on the number of files BEFORE we start messing with those files. But because some threads will still be enumerating dirs while others are getting file info, what do you do about the progress bar?

My initial thought was to have two stages of progress. The first stage wont have progress at all (just a progressbar in the busy state) while the total number of files is figured out, and then we switch to the progress state like normal. Again the problem is the threads. How do we know when the last thread has switched to files?

At this point I can easily get it to work but it will be ugly and cludgy with status variables that are checked and changed and bleh. The better idea is to separate the process into two completely. Spawn threads to discovery dirs and files, join those threads, and THEN spawn new threads to gather file info. So this code is probably going away but I decided to commit it anyway. There may still be some value in it.