IMDb Data Gathering
Implemented the imdb data gatherer. Realized I needed to add a few new items to the database at the same time. I also found out that saving raw data in an sqlite TEXT field doesn't work well. So now I'm pickling the data to save it. Loading is a pain because I have to eval the string, otherwise pickle complains about it not being a byte string. Not sure how to work around that.

Now that I have an example database I can start to implement the searching functions as well as the display functions.