Fancy search page
Not only was the search page missing some options, but it also looked ugly as hell. Now its all generated on the fly. This means I can freely update the database schema and not have to worry about changing this side of things too much. All the fields will update automatically.

Still have to do some managing of the type of field inside changedFieldType (i.e. for star widget).

Also cleaned up a lot of the element names, especially the layouts.