Basic searching implemented
The very beginnings of the search functionality is now in place and you do get results from it. It is still rough around the edges and we need to indicate the searching status by greying out certain things and having a progress indicator.

Decided to reform the main library page as its own widget so I could use it for the search results pages.