Offline searching
Implemented searching using the S3 dataset in an sqlite file. Be warned, the entire dataset after generation is 8.7GB! Also no longer storing data in pickled format. Fixed an issue where the imdb thread was actually running in the main thread instead of the background. Fixed the progress dialog so that the details pane doesn't clear when toggled.

Currently having trouble getting the correct data for the movie its searching for. There are so many results and many of them have the same year and title but are completely different. So far checking for things like cast members and directors seems to work.

Finally, many movies are being skipped because they have cryptic filenames that can't be parsed at all. We should instead use the folder name the file is in. Thats the next change.