Better matching and fixed search tab
The imdb search now can fall back to the folder name in case its a better title match than the file name. Search tabs are now closable and show the number of results in the tab title.

Having so many problems with the C3 database. My db isnt that old, was generated in nov of 2020, but its missing so much data. Most movies only have 3 or 4 cast members listed for instance. Now it uses the C3 database just for the initial search. Once its confident in a match it will pull data from imdb online to get all the data. This is more reliable but slower. Even worse is that there are differences between the formatting of the data between the online and c3 databases which is making errors pop up all over the place.