improved imdb search filtering and working star rating
Now checks our database against the C3 imdb title (in addition to the filename title) to prevent titles that are already in the database from being downloaded again from the online imdb.

Star rating widget now properly updates the star rating for each movie. If you click on the current star rating it will be set to 0 stars, allowing you to unrate a movie.