Updated scan progress bar
Made some changes to the library scan progress bar. It now shows the number of files processed in addition to the number of successful additions to the db. That means it also actually tracks the progress properly when theres a lot of dupes that get skipped over. The other change was to keep the width of the dialog when you press the details button. Previously it would reset the dialog's width to 500px.