Search and highlight update
Moved the searching into the library class where it should be. You can now include multiple values for a single field by separating values with a semicolon. Currently these all OR'ed together but in the future it will be selectable between AND'ing and OR'ing. The highlighting has been simplified and made more robust as I didn't know I could self-reference in a re.sub using \1. That makes a load of difference in the part of the code that keeps the case the same when replacing for a highlight. Also it now highlights character names that match.