Proper QThreads
I didn't do the threading correctly in any way, shape, or form. The main thread was still held up because my thread spawning was sleeping and never returning control to the main thread. Now the threads are spun up and then the function returns. Then as each thread finishes, it spins up a new thread using signals and slots. This works fantastically well.

Also beefed up the cancel button. Instead of waiting for mkvmerge to finish it just kills the processes manually. This is windows specific here, wont work on Linux although so far this has been developed for Windows.