Small bugfixes
Not opening the sub file as utf8 was causing major issues with some subs files so thats been fixed. The goodsubsids list is now generated through a for loop instead of list comprehension. This allows us to do additional checks without making the code look too hairy. Finally fixed a big in the mkextract command where the file path needed to be quoted but wasn't.