Dialog searching basically working
Got the dialog searching up and working. Although it only matches exact word for word quotes, it works surprisingly well. When a match is found it generates a VLC link that allows you to launch the movie at the point where that quote is said. The first few times it worked I just laughed, its so damn cool.

Next up will be to integrate these subtitles results into the search results page as well as implement more fuzzy searching.