Actually save star rating to db
Somehow forgot to call the update function in the movieLibrary class for the star rating. Probably because at the time I didn't have the movie title and wasn't in the mood to fix that. What's weird is nothing is every calling the update function but I do have multiple movies in my database with personal ratings on them, thats doing me a heckin big confuse. Did I call the update at some time and then break that code later on? This is why unit testing is a thing and I really should use it >.<.