Sort speedup
Not really speeding up the sorting itself but the process of rebuilding the list. Previously the QListWidget's signal for currentItemChanged() was still active and was firing for every single removal and add. As you can imagine this slows things down a lot so that signal is now disabled while the sort is happening and reenabled after the sort is complete.

Also added in a couple new sort options for the imdb rating and your personal ratings. It would be helpful in the future to show that rating when sorting by it, somewhere easy to see.