star rating bugfix
Fixed a crashed caused by changing the star rating widget when no movie is selected in the library or search view.

I also just discovered that the star rating widget isn't saving the ratings when you click from the main library view. And it crashes if you try to click the star rating widget in the search view. Not sure what's wrong there but for some reason findItems() is returning no results even when modified with a wildcard or single letter. Checking count() shows hundreds of items in the list but findItems() can't match any of them. Not my normal dev computer so that could have something to do with it.