tmp fix
Changing the library display to show the year of the film means the findItems() function now fails to find with Qt::MatchExact flag because the movie title doesnt have the year in it. For now we're modifying the title to search for by adding that date to it and this works for now... It breaks down when you sort by anything that changes the end of the movie title so for now we're matching with the Qt::MatchStartsWith flag and ignoring the end. This breaks for movies that have numbers at the end like the Toy Story series. Exact match solution involves recreating the title based on the sort mode as well.