Year sort bugfix
Sorting by year when you have also searched by year was causing a crash due to multiple factors. First it was trying to compare strings and ints which caused issues. This ultimately was also due to the year being replaced with a bunch of html code for the highlighting feature. Using the cleanyear key instead which isnt modified.