Dialog search timecode fix
I didn't realize that when I broke out of the loop I was left with the wrong value (value too big). Fixed by just holding onto the previous value and using that instead. This means we don't have to take 5 seconds off the timecode to line up properly anymore. Still a little delay because it feels better than no delay. Also added in the line previous in the text output. When that data is integrated into the GUI the specific quote will be highlighted as well.