Lots of crash fixes
Updated my local imdb s3 database and tried to rescan my library only to encounter a ton of problems. I fixed most of the crashes and an infinite loop I encountered and it seems mostly ok now. Also got a bunch of warnings about string escape sequences when trying this code in Python 3.12, so I changed all those strings with errors to raw strings. Theres some errors coming out of the imdbpy library but thats probably because its deprecated and is now Cinemagoer. Not really into the idea of updating everything to make use of the new library and it seems to still work ok for now. It does give some scary lookking IMDbDataAccessErrors but it  returns good data.