Color indicator for subtitles
Any movie titles with subtitle data now has a slight green background to indicate that in the main library view. Sometimes I search for a piece of dialog I know is correct and the movie is in my library, but no results. I can only guess if there is even subtitle data without opening the sqlite database manually and checking. So now I can just look.

One thing I would like to do is add a way to manually add subtitles. Some sort of right-click menu and dialog to allow pasting the subtitles or possibly a file browser to load an srt or something.