No subs indicator swap
Swapping around the indicator to show which titles lack subtitles as opposed to showing the titles that did have them. Using a faint red shade to color the background of any title missing subtitle data.

Also fixed a bug (I think). It checks two different fields for a good movie title in the hopes that its more resilient to weirdness by having two different sources to choose from (file name and folder name, should be the same but sometimes different). But the code was written in a way where both titles would have to be good for the imdb data to be accepted. I've swapped that and to an or and I think that makes sense. Now it will succeed if either is correct.

Next up will be adding a right click menu to the listwidget view with the option to edit the subtitle data. Might even be a good idea to add an option to edit the title data as well.