Right click menu start
Started work on the right click menu. Currently the context menu pops up and you can select edit, but it just prints the current movie title to console. Now to design the interface for that edit dialog. I think to start it will be the easy mode, just a text box with nothing (or the subtitles for the movie if it exists already in the database) and some save or cancel buttons. Also didnt realize there was a UI file for the movielibraryinfowidget. I guess I used that as a starting point but its not usable to create that widget without heavy modification. I still added the context menu change to the UI file in case I ever needed to reacreate it in the future.