Edit subtitles dialog v0.1
Initial work on the subtitles dialog. Was going to finish it all in one go but I realized the subtitles database needs a lot of data besides the movie title and subtitles, data I don't have ready to supply. So I'll have to add code to query that data from the normal title database so I can use that for the subtitle database.

At any rate, the dialog UI is done and it will pop up when you click on "Edit subtitles" after right clicking on any title. It also loads up the current subtitles if any exist in the database for the selected title.

Had to add a method to the subtitles database code for getting a single movie's subtitles. Only other way was to "search" for it with the movie title. Weird that I never wrote such a simple function but I never needed it until now.

All thats left is to simply query the needed extra data from the normal title database and use that when saving the subtitle data.