Finished sub edit dialog and annoying bugfix
I didn't realize I actually had all the data I needed already, it was in the userdata I was already using to get the movietitle. So it was a lot easier to get finished than I thought. It now allows you to edit and save any movie's subtitles. Now the sub library can be updated without having to scan each file manually to extract the subs. That took ages even on fast hardware and networks. After saving the subtitles a popup will tell you it succeeded and how much data was saved (number of chars).

The massively annoying bug fix is something I've been struggling with since its inception, and that was the link to play the movie at the found quote. It works great on titles with just letters, but any title that has an apostrophe or quotes would cause Qt to cut off the url at that point instead of using the full link. The solution was right in front of me the entire time, I was already decoding the string because Qt was encoding it on saving (which is probably when it was getting cut off). Instead of saving the raw string, now its first encoded manually with urllib. That hides the single or double quotes from Qt so nothing gets cut off. Its then decoded the same as usual and we now have the full URL. Wish I had thought of it sooner!

Also I have no idea what I was planning on doing with the subsdict{} inside subtitles.py but right now it doesnt do anything but throw exceptions every now and then. So its commented out until I figure out what I was thinking there.