Add delete option to new right click menu and bugfix
Added the ability to permanently delete a movie from the library and subtitle databases through the right click menu in the main library list. There is a warning popup that makes sure the user knows this is a permanent deletion.

Found a bug where I didn't realize I was getting a reference to the UserRole data of the QListWidgetItem and not a copy as I thought. Made that a deepcopy to prevent any funny business.

Also removed unnecessary properties from the flags set for the normal listwidget item in editableListWidget's addNormalItem() function. Also found a bug here where some films have actors in their list with no character, so instead of a dict with actor name and character name we get a string instead and that breaks things. Should be a simple fix, for later.