Added new button to movie data edit dialog
There are instances where you wouldnt want to add a character name to an actor list and would rather use a normal person name entry. Now the actor section of the movie data editing dialog has an extra blue button to add a character instead of a normal name. Also fixed a bug related to movies that had a mix of character/actor and just actor names in the actor field. Another fixed bug was the dialog accepting when you pressed enter, like after entering text in to a textbox and pressing enter to set the text. Now it only accept the dialog if you press save.