Finished edit movie data dialog
Added in the final code for the edit movie data dialog to retrieve the data from all the different widgets, organize them in an OrderedDict, and then save them to the database and update the library display with the new data. Seems to save data the same way it came in, so I think Im done.