Search sort bugfix
Adding the subtitle check to the sort code borked the search window pane since I wasnt setting the subtitle library reference there properly. Probably should require that when the object is initialized. Also realized that checking every item in the widgetlist for subtitles in the db each time its sorted is a bad idea. Think I'll change the subtitle adding code to set a value True in one of the extra fields of the movie data db. Will have to update the movie db with that value for every title currently with subtitles. That should speed things up considerably.