More sort search bugfixes
Didnt notice this until after finishing the last bit, the sorting by year is off. It was treating the year as a string and because it had a bunch of HTML tags in it for dumb reasons (highlighting it in the results but year is never highlighted anyway so its dumb and unnecessary) it wasn't putting them in actual numerical order (the 80's were after the 90's and before the 2000's for example). I didn't bother fixing that dumb code and just added a check to see if we needed to use the cleanyear instead. That whole cleantitle/cleanyear code suuuuuuuucks. But I'm too lazy to rewrite it.