Add movie dialog start work
Starting the process of adding a dialog to manually add a movie to the database. Was going to reuse the edit movie data dialog entirely but realized that would suuuuuuuck having to enter every single name and character one by one. So the idea for now is a simple dialog, file path and IMDB ID. Then if we get good data from that, and the movie title isnt already in the database, we can take the data from our local IMDB db and populate an edit movie data dialog with that data. That way the user can see what exactly is being saved and maybe edit anything before its saved. For now I just have the first simple dialog done.