Roadblock encountered for manual add dialog
Got it mostly working only to discover a couple things. One, theres a dumb bug in the current code that adds the IMDB data to the database where it converts lists into strings. Its fine since it runs an eval on the string to load it, but its unsightly before it gets done. Fixed that bug in the new section (havent fixed the old bug yet). But now I find out the IMDB database I have SUCKS! Its got NOTHING in it, barely a fraction of the information in the full database. It doesnt even have the runtime for new movies! And for old movies, like for instance Close Encounters of the Third Kind, there are only 4 actors listed for the entire movie! 4! What a fucking nightmare.

The only real way forward is something I REALLY dont want to do. I have to convert from IMDBpy to Cinamagoer. Have you seen the shittastic code that is library_scanner.py?? I dont even want to look at that code, much less change ALL of it. Shit balls. Gonna think about this for a day. Maybe I can shoe-horn in the new version without wrecking the old stuff. But since the old stuff is useless anyway whats the point? It needs to be redone. I just don't want to do it!