Case insensitive title sorting
Realized that movie titles are sorted in a case-sensitive way (e.g. TRON Legacy would be waaaay at the beginning of the T's with the original Tron movie down at the end of the T's). Just added a simple .lower() to our sortfunc code. Although now that I look at that code, it looks like that will also make sorting by other string keys case insensitive, which is good anyway. Had to swap the order of running loadLibraryIntoGui() and guimods() - running the guimods() first to assign the sublibref that the sorting function needs to color everything correctly. But now I realize that it was still coloring missing sub titles - so either that code is doubled somewhere or it WAS running the sort at start but ignoring the title sort key and just coloring. Weird.