Fix some scanning code
Finally went over that bit of code I was unsure about and just removed the second check on title2 entirely, not needed, it never gets used. After removing that code a few more titles were scanned into the library, although it did dupe the godfather again since theres a difference in databases with some titles having a colon and some not.

ALSO, Tried using the manual add dialog and it was so bad. It barely had any of the correct data. And when I add using the scanner, it always gives a redirect error when it tries to grab the online info - BUT it seems to actually retrieve good data anyway. I always thought that code was dead and not working. Anyway I think if I splice that code into the add movie dialog, it will actually work! Also still have that bug where the data is saved for the very first time, is displayed in its raw data format (dicts, lists, etc). Only after closing and reopening does the data get loaded in a human-readable format. Something to work on.