Manual add dialog working
After I realized the online imdb code actually did work, I incorporated that into the manual add dialog and now we're getting a full set of data! Saving that data was super easy, and updating the library display was just a single callback function. So barring any unforeseen bugs, I think the add movie dialog is finished!

Figured out the bug where scanning for new movies would add movies to the library view that had their data in raw form (e.g. list of directors would be a raw python list ['director one', 'directory two', 'etc']. For some reason I was str()'ing all the lists and stuff before saving to the sqlite db. When we save the new movie data from the scanner, it saves it into both the sqlite database and a dictionary thats internal to the movie library class object. When a full library refresh is requested it checks if self.librarydict is not None and then returns that dict instead of the data from the library to speed things up. This only happens when a full library is requested, which is what we do (instead of manually adding the new titles to the library display ourselves). At first start everything is fine because there is no librarydict, so it pulls the data from the db and instantiates it as whatever type it needs to be.

Still not sure why we had those str() calls around all the objects, but I think it was trying to fix a bug where selecting a newly added title in the library view just after scanning and adding it, the app would crash. That bug is gone now for whatever reason, so its not a problem.